Set KUBECONFIG to the $HOME/.kube/config if missing
To support root or non-root docker invocations, set the KUBECONFIG
specifically to the passed .kube dir. The reason is that 'oc' incovation
as root is by default looking for the .kube dir uner
`/root/.kube/config` where as non-root is looking at $HOME/.kube/config.

With this we always set KUBECONFIG, so root invocation of `oc` commands
will always find the passed in config.

Fixes: https://github.com/ansibleplaybookbundle/ansible-playbook-bundle/issues/298

Signed-off-by: Roy Golan <rgolan@redhat.com>